Move low level function that hit HTTP APIs into a separate package
