+++ pkg/replicatedapp/upstream_test.go
@@ -1,0 +1,66 @@
+package replicatedapp
+
+import (
+	"net/url"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+var (
+	v1_2_0 = "v1.2.0"
+)
+
+func Test_parseReplicatedURL(t *testing.T) {
+	tests := []struct {
+		name                 string
+		uri                  string
+		expectedAppSlug      string
+		expectedChannel      *string
+		expectedVersionLabel *string
+		expectedSequence     *int
+	}{
+		{
+			name:                 "replicated://app-slug",
+			uri:                  "replicated://app-slug",
+			expectedAppSlug:      "app-slug",
+			expectedChannel:      nil,
+			expectedVersionLabel: nil,
+			expectedSequence:     nil,
+		},
+		{
+			name:                 "replicated://app-slug@v1.2.0",
+			uri:                  "replicated://app-slug@v1.2.0",
+			expectedAppSlug:      "app-slug",
+			expectedChannel:      nil,
+			expectedVersionLabel: &v1_2_0,
+			expectedSequence:     nil,
+		},
+		{
+			name:                 "replicated://app-slug/channel",
+			uri:                  "replicated://app-slug/channel",
+			expectedAppSlug:      "app-slug",
+			expectedChannel:      &channel,
+			expectedVersionLabel: nil,
+			expectedSequence:     nil,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			req := require.New(t)
+
+			u, err := url.ParseRequestURI(test.uri)
+			req.NoError(err)
+
+			replicatedUpstream, err := ParseReplicatedURL(u)
+			req.NoError(err)
+			assert.Equal(t, test.expectedAppSlug, replicatedUpstream.AppSlug)
+
+			if test.expectedVersionLabel != nil || replicatedUpstream.VersionLabel != nil {
+				assert.Equal(t, test.expectedVersionLabel, replicatedUpstream.VersionLabel)
+			}
+		})
+	}
+}
